edit sign up and sign in

+++ src/pages/SignIn.js
@@ -5,6 +5,7 @@ import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
+import {Link as RouterLink} from "react-router-dom";
 import Link from "@mui/material/Link";
 import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
@@ -12,47 +13,42 @@ import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
 import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
-import {auth, provider} from '../firebase-config';
+import {auth, provider, db} from '../firebase-config';
 import {signInWithPopup} from 'firebase/auth';
 import { useNavigate } from "react-router-dom";
+import { doc, getDoc, setDoc } from "firebase/firestore"; 
+
 
 const theme = createTheme();
 
-export default function SignIn({setIsAuth}) {
+const USERS = 'users';
+
+export default function SignIn(props) {
   let navigate = useNavigate();
 
   const signInWithGoogle = () => {
     signInWithPopup(auth, provider).then((result) => {
-      localStorage.setItem('isAuth', true);
-      console.log('isauth');
-      setIsAuth(true);
-      navigate('/');
+      props.signUserUp(result.user.displayName, result.user.email);
     })
+    localStorage.setItem('isAuth', true);
+    console.log('isauth');
+    props.setIsAuth(true);
+    navigate('/');
   };
 
-  const handleSubmit = async (event) => {
+  const signIn = (event) => {
     event.preventDefault();
-    const data = new FormData(event.currentTarget);
-    // eslint-disable-next-line no-console
-    console.log({
-      email: data.get("email"),
-      password: data.get("password"),
-    });
-    const requestOptions = {
-      method: "POST",
-      mode: "no-cors",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        email: data.get("email"),
-        password: data.get("password"),
-      }),
-    };
-    let response = await fetch("/api", requestOptions);
-    //response = await response.json();
-
-    //this.setState({ postId: data.id });
+    const userDocRef = doc(db, USERS, event.target.userName.value);
+    getDoc(userDocRef).then((userDocSnap) => {
+      if(userDocSnap.exists() && userDocSnap.data().password === event.target.password.value){
+        localStorage.setItem('isAuth', true);
+        console.log('isauth');
+        props.setIsAuth(true);
+        navigate('/');
+      }
+      else
+        console.log('username or password are not correct');
+    })
   };
 
   return (
@@ -75,7 +71,7 @@ export default function SignIn({setIsAuth}) {
           </Typography>
           <Box
             component="form"
-            onSubmit={handleSubmit}
+            onSubmit={signIn}
             noValidate
             sx={{ mt: 1 }}
           >
@@ -83,10 +79,10 @@ export default function SignIn({setIsAuth}) {
               margin="normal"
               required
               fullWidth
-              id="email"
-              label="Email Address"
-              name="email"
-              autoComplete="email"
+              id="username"
+              label="User Name"
+              name="userName"
+              autoComplete="username"
               autoFocus
             />
             <TextField
@@ -99,10 +95,6 @@ export default function SignIn({setIsAuth}) {
               id="password"
               autoComplete="current-password"
             />
-            <FormControlLabel
-              control={<Checkbox value="remember" color="primary" />}
-              label="Remember me"
-            />
             <Button
               type="submit"
               fullWidth
@@ -113,21 +105,18 @@ export default function SignIn({setIsAuth}) {
             </Button>
             <Grid container>
               <Grid item xs>
-                <Link href="#" variant="body2">
-                  Forgot password?
-                </Link>
+                <RouterLink to='#'>
+                  <Link variant="body2">
+                    Forgot password?
+                  </Link>
+                </RouterLink>
               </Grid>
               <Grid item>
-                <Link
-                  href="#"
-                  variant="body2"
-                  onClick={() => {
-                    sessionStorage.setItem("signup", 1);
-                    window.location.reload();
-                  }}
-                >
-                  {"Don't have an account? Sign Up"}
-                </Link>
+                <RouterLink to='/signUp'>
+                  <Link variant="body2">
+                    {"Don't have an account? Sign Up"}
+                  </Link>
+                </RouterLink>
               </Grid>
             </Grid>
             <button className="signin-with-google-btn" onClick={signInWithGoogle}>Sign in with google</button>
